Fix imports after move
